almost finish stream_inference
