+++ requirements.txt
@@ -9,3 +9,5 @@ sentencepiece
 transformers @ git+https://github.com/huggingface/transformers@c612628045822f909020f7eb6784c79700813eda
 flask
 flask_cors
+icetk==0.0.7
+cpm_kernels==1.0.11
