+++ src/lmflow/models/auto_model.py
@@ -5,7 +5,7 @@
 
 from lmflow.models.hf_decoder_model import HFDecoderModel
 from lmflow.models.text_regression_model import TextRegressionModel
-
+from lmflow.models.hf_encoder_decoder_model import HFEncoderDecoderModel
 
 class AutoModel:
 
@@ -16,6 +16,8 @@ class AutoModel:
             return HFDecoderModel(model_args, *args, **kwargs)
         elif arch_type == "text_regression":
             return TextRegressionModel(model_args, *args, **kwargs)
+        elif arch_type == "encoder_decoder":
+            return HFEncoderDecoderModel(model_args, *args, **kwargs)
         else:
             raise NotImplementedError(
                 f"model architecture type \"{arch_type}\" is not supported"
